+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - switched to underscores in project name
+- added placeholder support to: `to-mel-spectrogram`, `to-mfcc-spectrogram`, `to-stft-spectrogram`
 
 
 0.0.1 (2024-07-05)